# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GraphFlow",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/graph/generate": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Generate GraphData and Excalidraw scene from user content.",
        "operationId": "handle_generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Graph generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/graph/render": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Render a scene (or GraphData) to PNG/SVG artifacts.",
        "operationId": "handle_render",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rendered artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Container": {
        "type": "object",
        "required": [
          "id",
          "label",
          "children"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "style": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContainerStyle"
              }
            ],
            "nullable": true
          }
        }
      },
      "ContainerStyle": {
        "type": "object",
        "properties": {
          "bg": {
            "type": "string",
            "nullable": true
          },
          "border": {
            "type": "string",
            "nullable": true
          },
          "radius": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "label_tag": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Decoration": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "target": {
            "type": "string",
            "nullable": true
          },
          "at_x": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "at_y": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "builtin": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DecorationSize"
              }
            ],
            "nullable": true
          },
          "offset": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DecorationOffset"
              }
            ],
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "DecorationOffset": {
        "type": "object",
        "required": [
          "dx",
          "dy"
        ],
        "properties": {
          "dx": {
            "type": "number",
            "format": "float"
          },
          "dy": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "DecorationSize": {
        "type": "object",
        "required": [
          "w",
          "h"
        ],
        "properties": {
          "w": {
            "type": "number",
            "format": "float"
          },
          "h": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "EdgeData": {
        "type": "object",
        "required": [
          "id",
          "source",
          "target",
          "label",
          "style"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "style": {
            "$ref": "#/components/schemas/EdgeStyle"
          }
        }
      },
      "EdgeStyle": {
        "type": "object",
        "required": [
          "line",
          "arrow"
        ],
        "properties": {
          "line": {
            "type": "string"
          },
          "arrow": {
            "type": "string"
          }
        }
      },
      "GenerateRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "tier": {
            "type": "string",
            "nullable": true
          },
          "allow_images": {
            "type": "boolean",
            "nullable": true
          },
          "assets_dir": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GenerateResponse": {
        "type": "object",
        "required": [
          "graph_data",
          "scene",
          "artifacts"
        ],
        "properties": {
          "graph_data": {
            "$ref": "#/components/schemas/GraphData"
          },
          "scene": {},
          "artifacts": {}
        }
      },
      "GlobalStyle": {
        "type": "object",
        "required": [
          "font",
          "background"
        ],
        "properties": {
          "font": {
            "type": "string"
          },
          "background": {
            "type": "string"
          },
          "theme": {
            "type": "string",
            "example": "minimal",
            "nullable": true
          }
        }
      },
      "GraphData": {
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeData"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EdgeData"
            }
          },
          "layout_hints": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LayoutHints"
              }
            ],
            "nullable": true
          },
          "global_style": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GlobalStyle"
              }
            ],
            "nullable": true
          },
          "decorations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Decoration"
            },
            "nullable": true
          },
          "containers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Container"
            },
            "nullable": true
          }
        }
      },
      "LayoutHints": {
        "type": "object",
        "required": [
          "direction",
          "algorithm"
        ],
        "properties": {
          "direction": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          }
        }
      },
      "NodeData": {
        "type": "object",
        "required": [
          "id",
          "label",
          "x",
          "y",
          "style"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          },
          "style": {
            "$ref": "#/components/schemas/NodeStyle"
          }
        }
      },
      "NodeStyle": {
        "type": "object",
        "required": [
          "shape",
          "color"
        ],
        "properties": {
          "shape": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        }
      },
      "RenderRequest": {
        "type": "object",
        "properties": {
          "scene": {
            "nullable": true
          },
          "graph_data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GraphData"
              }
            ],
            "nullable": true
          },
          "filename_hint": {
            "type": "string",
            "nullable": true
          },
          "formats": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
      "RenderResponse": {
        "type": "object",
        "required": [
          "suggested"
        ],
        "properties": {
          "suggested": {
            "type": "string"
          },
          "png": {
            "type": "string",
            "nullable": true
          },
          "svg": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "graph",
      "description": "Graph generation and rendering APIs"
    }
  ]
}